Fxied Bug With UART Implimentation, Fixed bug in print() function, Fixed bug in Makefile, Modified Testing Code

+++ Firmware/uart.h
@@ -30,7 +30,7 @@
 
 // Functions
 
-void usart_init(uint32_t ubrr); // Initializes USART registers by specification and ubrr value. Author: Hao Lin (21/08/2020)
+void usart_init(uint16_t ubrr); // Initializes USART registers by specification and ubrr value. Author: Hao Lin (21/08/2020)
 void usart_transmit(uint8_t data); // Defining function to transmit 1 character at a time. Author: Krithik Lekinwala (22/08/2020)
 void usart_transmitRaw(uint8_t rawData[], size_t arraySize); // Defining function to transmit single raw bytes one at a time. Author: Krithik Lekinwala (22/08/2020)
 void print_array_intergers(uint16_t intArray[], uint16_t arrayLength); // Print out an array of POSITIVE integers. Inputs will be array of type int, and length of the array. Author: Hao Lin (22/08/2020)
